+++ changes.md
@@ -8,3 +8,6 @@ This is a log of where we have had to differ from the class diagram, storyboards
 *   Function int getId() added to UserAccount so that UserAccountManager can run deleteUserAccountById(int Id)
 *   Function List<Token> getTokens() added to UserAccount so that tokens can be matched to a user.
 *   Function String getName() added to Person so that a UserAccount can be fount by name using UserAccountManager.
+*   Function List<Pass> getAllPasses added to UserAccount so that UserAccountManager can deactivate invalid passes.
+*   Function boolean isValid() added to PassPlan so that it can check weather the PassPlan is valid;
+*   Function PassPlan getPassPlan() added to Pass as to check it is valid you need to access the PassPlan